@@ -6,9 +6,24 @@
   "scripts": {
     "precommit": "./scripts/pre_commit_hook.sh",
     "format": "prettier --no-semi --single-quote --bracket-spacing --braces-spacing --jsx-single-quote --space-before-function-paren --write \"src/**/*.{ts,tsx}\"",
-    "start": "ts-node fuse",
-    "build": "NODE_ENV='production' ts-node fuse",
+    "bundle": "webpack",
+    "dev": "webpack-dev-server",
     "test": "jest"
+  },
+  "prettier": {
+    "printWidth": 80,
+    "tabWidth": 2,
+    "useTabs": false,
+    "semi": false,
+    "singleQuote": true,
+    "jsxSingleQuote": true,
+    "trailingComma": "none",
+    "bracketSpacing": true,
+    "bracesSpacing": true,
+    "jsxBracketSpacingSameLine": false,
+    "alignObjectProperties": true,
+    "noSpaceEmptyFn": false,
+    "spaceBeforeFunctionParen": true
   },
   "jest": {
     "transform": {
@@ -33,7 +48,6 @@
   },
   "dependencies": {
     "csstips": "^0.2.0",
-    "husky": "^0.14.2",
     "lodash": "^4.17.4",
     "material-ui": "^0.18.5",
     "mobx": "^3.1.16",
@@ -41,8 +55,7 @@
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
     "react-tap-event-plugin": "^2.0.1",
-    "typestyle": "^1.3.7",
-    "webpack": "^3.0.0"
+    "typestyle": "^1.3.7"
   },
   "devDependencies": {
     "@types/jest": "^19.2.3",
@@ -52,12 +65,19 @@
     "@types/react": "^15.0.34",
     "@types/react-dom": "^15.5.0",
     "@types/react-tap-event-plugin": "^0.0.30",
-    "@types/webpack": "^3.0.1",
+    "@types/webpack": "^3.0.4",
+    "awesome-typescript-loader": "^3.2.1",
     "coveralls": "^2.13.1",
+    "fork-ts-checker-notifier-webpack-plugin": "^0.2.0",
+    "fork-ts-checker-webpack-plugin": "^0.2.7",
+    "husky": "^0.14.2",
     "jest": "^20.0.3",
-    "prettier-miscellaneous": "^1.5.2-fix",
+    "prettier-miscellaneous": "^1.5.2",
     "ts-jest": "^20.0.4",
-    "ts-node": "^3.0.4",
-    "typescript": "^2.3.3"
+    "ts-loader": "^2.3.0",
+    "ts-node": "^3.2.0",
+    "typescript": "^2.3.3",
+    "webpack": "^3.3.0",
+    "webpack-dev-server": "^2.5.1"
   }
 }
